initial move of type policy generation
