+++ src/packages/admin-ui-components/src/utils/use-schema.ts
@@ -1,5 +1,7 @@
 import { useMemo } from 'react';
 import { ApolloCache, TypePolicy, useQuery } from '@apollo/client';
+import { generateTypePolicies } from '@exogee/graphweaver-apollo-client';
+
 import { SCHEMA_QUERY } from './graphql';
 import { PanelMode } from '../detail-panel';
 
@@ -104,43 +106,12 @@ export interface SortField {
 	direction: SortDirection;
 }
 
-type Cache = ApolloCache<unknown> & {
-	policies: { addTypePolicies: (policy: { Query: TypePolicy }) => void };
-};
-
 type EntityMap = {
 	[entityName: string]: Entity;
 };
 
-const generateTypePolicyFields = (entityMap: EntityMap) => {
-	const policy = {
-		keyArgs: (
-			args: {
-				filter?: Record<string, unknown>;
-				pagination?: { orderBy: Record<string, unknown> };
-			} | null
-		) => {
-			// https://www.apollographql.com/docs/react/pagination/key-args/#keyargs-function-advanced
-			const filter = args?.filter ? JSON.stringify(args.filter) : '';
-			const orderBy = args?.pagination?.orderBy ? JSON.stringify(args.pagination.orderBy) : '';
-			return btoa(`${filter}:${orderBy}`);
-		},
-		merge(existing = [], incoming: { __ref: string }[]) {
-			const merged = [...existing, ...incoming];
-			const uniqueItems = new Set(merged.map((item) => item.__ref));
-			return [...uniqueItems].map((__ref) => merged.find((item) => item.__ref === __ref));
-		},
-	};
-
-	const mapEntityToPolicy = (entity: Entity) => ({
-		[entity.plural.charAt(0).toLowerCase() + entity.plural.slice(1)]: policy,
-	});
-
-	return {
-		...Object.values(entityMap)
-			.map(mapEntityToPolicy)
-			.reduce((acc, policy) => ({ ...acc, ...policy }), {}),
-	};
+type Cache = ApolloCache<unknown> & {
+	policies: { addTypePolicies: (policy: { Query: TypePolicy }) => void };
 };
 
 export const useSchema = () => {
@@ -164,7 +135,7 @@ export const useSchema = () => {
 
 	// We already have an array of entities but we should pre-build a lookup by name.
 	const entityMap = useMemo(() => {
-		const result: { [entityName: string]: Entity } = {};
+		const result: EntityMap = {};
 		if (!data?.result?.entities) return result;
 
 		for (const entity of data.result.entities) {
@@ -172,13 +143,9 @@ export const useSchema = () => {
 		}
 
 		// Now we have our entities we can create the type policy
-		const typePolicy: { Query: TypePolicy } = {
-			Query: {
-				keyFields: ['id'], // This is the default and is here for clarity
-				fields: generateTypePolicyFields(result),
-			},
-		};
-		cache.policies.addTypePolicies(typePolicy);
+		const entityNames = data.result.entities.map((entity) => entity.plural);
+		const typePolicies = generateTypePolicies(entityNames);
+		cache.policies.addTypePolicies(typePolicies);
 
 		return result;
 	}, [data]);
@@ -198,15 +165,12 @@ export const useSchema = () => {
 		error,
 		entities: Object.keys(entityMap),
 		backends: Object.keys(dataSourceMap),
-		entityByName: (entityName: string) => entityMap[entityName],
+		entityByName: (entityName: string) => entityMap[entityName as keyof typeof entityMap],
 		entityByType: (entityType: string) => {
 			const entityName = entityType.replaceAll(/[^a-zA-Z\d]/g, '');
 			return entityMap[entityName];
 		},
 		enumByName: (enumName: string) => enumMap[enumName],
 		entitiesForBackend: (backendId: string) => dataSourceMap[backendId],
-		entityInBackend: (entityName: string, backendId: string) =>
-			// TODO: This could be an O(1) lookup if we build one first.
-			!!dataSourceMap[backendId].find((entity) => entity.name === entityName),
 	};
 };
