+++ src/packages/apollo-client/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+	ignorePatterns: ['.deploy/**', '.build/**', 'lib/**'],
+};
