+++ src/packages/apollo-client/.npmrc
@@ -1,0 +1,1 @@
+save-exact=true
