+++ src/packages/apollo-client/README.md
@@ -1,0 +1,13 @@
+# `@exogee/graphweaver-apollo-client`
+
+Useful helpers for working with Apollo Client and Graphweaver.
+
+
+
+## Documentation
+
+Comprehensive documentation and usage examples can be found on our [Docs Site](https://graphweaver.com/docs). It covers installation instructions, detailed API documentation, and guides to help you get started with Graphweaver.
+
+## Graphweaver CLI `graphweaver`
+
+The Graphweaver Command Line Interface (CLI) tool enables you to set up and manage Graphweaver using commands in your command-line shell. Check the `graphweaver` npm package [here.](https://www.npmjs.com/package/graphweaver)
