+++ src/packages/apollo-client/build.js
@@ -1,0 +1,9 @@
+import esbuild from 'esbuild';
+(async () => {
+	await esbuild.build({
+		outdir: 'lib',
+		format: 'esm',
+		sourcemap: 'linked',
+		entryPoints: ['src/index.ts'],
+	});
+})();
